requirements.txt
aws-cdk-lib==2.185.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v32==2.1.0
